# docs/swagger.yaml
definitions:
  dto.CourseDto:
    properties:
      detail:
        type: string
      image:
        type: string
      name:
        type: string
      step:
        items:
          $ref: '#/definitions/dto.CourseStepDto'
        type: array
      tags:
        example:
        - id1
        - id2
        items:
          type: string
        type: array
    type: object
  dto.CourseStepDto:
    properties:
      content:
        type: string
      order:
        type: integer
      second:
        type: integer
    type: object
  dto.UserInfoDto:
    properties:
      birthday:
        type: integer
      gender:
        type: integer
      nickName:
        example: 傻逼
        type: string
    type: object
  dto.UserLoginDto:
    properties:
      identityToken:
        example: 你他妈向苹果登录那个token
        type: string
    type: object
  response.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact:
    name: orangeboyChen
  description: 如有问题，请联系orangeboy
  title: 不叫外卖后端API文档
  version: "1.0"
paths:
  /course:
    post:
      parameters:
      - description: 菜品详情
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CourseDto'
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 上传菜品
      tags:
      - 菜品
  /course/{courseId}:
    delete:
      parameters:
      - description: courseId
        in: path
        name: courseId
        required: true
        type: string
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 删除菜品
      tags:
      - 菜品
    get:
      parameters:
      - description: 菜品id
        in: path
        name: courseId
        required: true
        type: string
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 获取菜品详情
      tags:
      - 菜品
    put:
      parameters:
      - description: courseId
        in: path
        name: courseId
        required: true
        type: string
      - description: 菜品详情
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CourseDto'
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 更新菜品
      tags:
      - 菜品
  /course/query:
    get:
      parameters:
      - description: '查找类型，可选: tag'
        in: query
        name: by
        required: true
        type: string
      - description: 关键词
        in: query
        name: keyword
        required: true
        type: string
      - description: 页数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页面大小
        in: query
        name: pageSize
        required: true
        type: integer
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 查找菜品
      tags:
      - 菜品
  /course/recommend:
    get:
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 获取推荐列表
      tags:
      - 菜品
  /course/search:
    get:
      description: 根据关键字搜索菜品
      parameters:
      - description: 关键词
        in: query
        name: keyword
        required: true
        type: string
      - description: 页数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页面大小
        in: query
        name: pageSize
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
      security:
      - ApiAuthToken: []
      summary: 搜索菜品
      tags:
      - 菜品
  /login:
    post:
      parameters:
      - description: 登录数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Result'
      summary: 登录
      tags:
      - 鉴权
  /tag/type/{tagTypeId}:
    get:
      parameters:
      - description: tagTypeId
        in: path
        name: typeId
        required: true
        type: string
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 获取标签列表
      tags:
      - 标签
  /tag/type/list:
    get:
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 获取标签种类列表
      tags:
      - 标签
  /user/avatar:
    put:
      parameters:
      - description: 头像
        in: formData
        name: avatar
        required: true
        type: file
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 上传头像
      tags:
      - 用户
  /user/avatar/{avatarFileName}:
    get:
      parameters:
      - description: avatarFileName
        in: path
        name: avatarFileName
        required: true
        type: string
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 获取用户头像
      tags:
      - 用户
  /user/info:
    put:
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserInfoDto'
      responses: {}
      security:
      - ApiAuthToken: []
      summary: 更新用户信息
      tags:
      - 用户
securityDefinitions:
  ApiAuthToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"